Split SocketModeClient.ts into several files
