+++ packages/socket-mode/src/index.ts
@@ -1,8 +1,8 @@
 /// <reference lib="es2017" />
 
-export {
-  SocketModeClient,
-} from './SocketModeClient';
+export { SocketModeClient } from './SocketModeClient';
+export { SocketModeOptions } from './SocketModeOptions';
+export { UnrecoverableSocketModeStartError } from './UnrecoverableSocketModeStartError';
 
 export { Logger, LogLevel } from './logger';
